requirements.txt
unittest2
pillow
numpy
opencv-python
requests
beautifulsoup4
pytesseract
begins